+++ src/game.cpp
@@ -173,7 +173,7 @@ void Game::set_current_frame_input(int keystate, int x, int y, int mousepress)
 		stringstream inputMsg;
 		string msg;
 
-		unsigned long checksum = world->checksum();
+		unsigned long long checksum = world->checksum();
 		inputMsg << "1 " << myID << " " << frame << " " << keystate << " " << x << " " << y << " " << mousepress << " " << checksum << "#";
 		msg = inputMsg.str();
 		clientSocket.write(SERVER_ID, msg);
