+++ src/world.h
@@ -70,7 +70,7 @@ public:
 	
 	void terminate();
 	
-	unsigned long checksum() const;
+	unsigned long long checksum() const;
 
 	void add_message(const std::string& message) const;
 	void add_event(const WorldEvent& event) const;
